Move frontend tests into dedicated directory
