+++ front-end/src/__tests__/utils/rosterData.test.js
@@ -9,7 +9,7 @@ import {
   getRosterPlayerAvatar,
   normalizeRosterName,
   ROSTER_SEASONS,
-} from './rosterData';
+} from '../../utils/rosterData';
 
 test('exposes roster seasons from 2025-26 through 2022-23', () => {
   expect(ROSTER_SEASONS.map(season => season.value)).toEqual([
